gerenciamento com Roles feito

+++ backend/src/business/PollsBusiness.ts
@@ -4,12 +4,15 @@ import { InputPollDTO, PollsModel } from "../models/PollsModel";
 import { Authenticator } from "../services/Autheticator";
 import { IdGenerator } from "../services/IdGenerator";
 import { BaseError } from "../error/ErrorBase";
+import UserDatabse from "../data/UsersDatabase";
+import { UserRole } from "../models/UsersModel";
 
 export default class PollsBusiness {
   constructor(
     private idGenerator = new IdGenerator(),
     private pollsData = new PollsDatabase(),
-    private authenticator = new Authenticator()
+    private authenticator = new Authenticator(),
+    private userData = new UserDatabse()
   ) {}
 
   public async createPoll(input: InputPollDTO, token: string) {
@@ -31,7 +34,7 @@ export default class PollsBusiness {
 
       const findTitle = await this.pollsData.findByTitle(title);
       if (findTitle) {
-        throw new BaseError(401,"This poll already exists");
+        throw new BaseError(401, "This poll already exists");
       }
 
       const [dayStart, monthStart, yearStart] = start_date.split("/");
@@ -56,6 +59,7 @@ export default class PollsBusiness {
 
       const newPoll = new PollsModel(
         id,
+        validToken.id,
         title,
         start_dateFormat,
         end_dateFormat
@@ -73,53 +77,96 @@ export default class PollsBusiness {
       let { start_date, end_date } = input;
 
       if (!token) {
-        throw new BaseError(404,"Token not found, please check login");
+        throw new BaseError(404, "Token not found, please check login");
       }
       const validToken = this.authenticator.getData(token);
-      if (!validToken) {
+      if (!validToken.id) {
         throw new BaseError(401, "Invalid token, check login");
       }
-
       const validPoll = await this.pollsData.findByID(id);
       if (!validPoll) {
         throw new BaseError(404, "Poll not found");
       }
 
-      const newInput: InputEditPollBDDTO = {
-        title,
-      };
+      const user = await this.userData.findByID(validToken.id);
+      const ceratorId = validPoll.getCreatorId();
 
-      if (start_date) {
-        const [dayStart, monthStart, yearStart] = start_date.split("/");
-        const start_dateFormat = new Date(
-          `${yearStart}-${monthStart}-${dayStart}`
-        );
+      if (user.getRole() === UserRole.ADMIN) {
+        const newInput: InputEditPollBDDTO = {
+          title,
+        };
 
-        if (
-          start_dateFormat.setUTCHours(0, 0, 0, 0) <
-          new Date().setUTCHours(0, 0, 0, 0)
-        ) {
-          throw new BaseError(400, "Invalid date");
+        if (start_date) {
+          const [dayStart, monthStart, yearStart] = start_date.split("/");
+          const start_dateFormat = new Date(
+            `${yearStart}-${monthStart}-${dayStart}`
+          );
+
+          if (
+            start_dateFormat.setUTCHours(0, 0, 0, 0) <
+            new Date().setUTCHours(0, 0, 0, 0)
+          ) {
+            throw new BaseError(400, "Invalid date");
+          }
+          newInput.start_date = start_dateFormat;
         }
-        newInput.start_date = start_dateFormat;
-      }
 
-      if (end_date) {
-        const [dayStart, monthStart, yearStart] = end_date.split("/");
-        const end_dateFormat = new Date(
-          `${yearStart}-${monthStart}-${dayStart}`
-        );
+        if (end_date) {
+          const [dayStart, monthStart, yearStart] = end_date.split("/");
+          const end_dateFormat = new Date(
+            `${yearStart}-${monthStart}-${dayStart}`
+          );
 
-        if (
-          end_dateFormat.setUTCHours(0, 0, 0, 0) <
-          new Date().setUTCHours(0, 0, 0, 0)
-        ) {
-          throw new BaseError(400, "Invalid date");
+          if (
+            end_dateFormat.setUTCHours(0, 0, 0, 0) <
+            new Date().setUTCHours(0, 0, 0, 0)
+          ) {
+            throw new BaseError(400, "Invalid date");
+          }
+          newInput.end_date = end_dateFormat;
         }
-        newInput.end_date = end_dateFormat;
-      }
 
-      await this.pollsData.editPoll(newInput, id);
+        await this.pollsData.editPoll(newInput, id);
+      } else {
+        if (user.getId() !== validPoll.getCreatorId()) {
+          throw new BaseError(403, "You can't edit others Polls");
+        }
+        const newInput: InputEditPollBDDTO = {
+          title,
+        };
+
+        if (start_date) {
+          const [dayStart, monthStart, yearStart] = start_date.split("/");
+          const start_dateFormat = new Date(
+            `${yearStart}-${monthStart}-${dayStart}`
+          );
+
+          if (
+            start_dateFormat.setUTCHours(0, 0, 0, 0) <
+            new Date().setUTCHours(0, 0, 0, 0)
+          ) {
+            throw new BaseError(400, "Invalid date");
+          }
+          newInput.start_date = start_dateFormat;
+        }
+
+        if (end_date) {
+          const [dayStart, monthStart, yearStart] = end_date.split("/");
+          const end_dateFormat = new Date(
+            `${yearStart}-${monthStart}-${dayStart}`
+          );
+
+          if (
+            end_dateFormat.setUTCHours(0, 0, 0, 0) <
+            new Date().setUTCHours(0, 0, 0, 0)
+          ) {
+            throw new BaseError(400, "Invalid date");
+          }
+          newInput.end_date = end_dateFormat;
+        }
+
+        await this.pollsData.editPoll(newInput, id);
+      }
     } catch (error: any) {
       throw new BaseError(400, error.message);
     }
@@ -141,7 +188,16 @@ export default class PollsBusiness {
         throw new BaseError(401, "Invalid toke, please check login");
       }
 
-      await this.pollsData.deletePoll(id);
+      const user = await this.userData.findByID(validToken.id)
+      if(user.getRole() === UserRole.ADMIN){
+        await this.pollsData.deletePoll(id);
+      }else{
+        if(user.getId() !== validId.getCreatorId()) {
+          throw new BaseError(403, "You can't delete others Poll")
+        }else{
+          await this.pollsData.deletePoll(id);
+        }
+      }
     } catch (error: any) {
       throw new BaseError(400, error.message);
     }
@@ -166,26 +222,26 @@ export default class PollsBusiness {
     }
   }
 
-  public async getPollById(id:string, token: string) {
+  public async getPollById(id: string, token: string) {
     try {
-      if(!token) {
-        throw new BaseError(422, "Token missing")
+      if (!token) {
+        throw new BaseError(422, "Token missing");
       }
-      if(!id){
-        throw new BaseError(422, "Please inform the Poll ID")
-      }
-
-      const validToken = this.authenticator.getData(token)
-      if(!validToken.id){
-        throw new BaseError(401, "Invalid token")
+      if (!id) {
+        throw new BaseError(422, "Please inform the Poll ID");
       }
 
-      const validPoll = await this.pollsData.findByID(id)
-      if(!validPoll){
-        throw new BaseError(404, "Poll not found")
+      const validToken = this.authenticator.getData(token);
+      if (!validToken.id) {
+        throw new BaseError(401, "Invalid token");
       }
 
-      return validPoll
+      const validPoll = await this.pollsData.findByID(id);
+      if (!validPoll) {
+        throw new BaseError(404, "Poll not found");
+      }
+
+      return validPoll;
     } catch (error: any) {
       throw new BaseError(400, error.message);
     }
